Update patch set 1
Patch Set 1:

I guess we could make it a script that checks for the existence of a file that is just a list of devices to enable it for, and follows that, or, if it doesn't exist, loops through all the devices enabling them.  I'm not opposed to that approach.

Patch-set: 1