Update patch set 14
Patch Set 14:

> This patch was only merged on the 21st, so it couldn't have caused
 > this problem

To be honest, I don't remember correctly the last working BUILD. As for High Touch Sensitivity option, I can't even go to Settings. The screen touch is messed up.

I am pretty sure that this patch is the cause of the problem (because Glove Mode does the same on Yureka PLUS but works fine on Yureka). 
By the way, is this HighTouch Sensitivity option enabled by default?

Patch-set: 14