Update patch set 14
Patch Set 14:

Alright, I figured out the Modes and also how to change them.
The display shows the same messed up behavior in both the Normal and Glove Mode.

So, it looks like that the Normal and Glove Mode switching is not working i.e., Glove Mode remains on irrespective of what Mode one selects. Thus, I think this is what needs to be fixed.

By the way, I don't think merging a feature which is partially working (works on Yureka but not on PLUS or vice-versa) is a wise move as the same source is used for both the devices. If you can put some kind of condition which checks what device the ROM is going to run on and then decides accordingly whether to run the feature or not, then it would be a win-win. However, if that can't be done, then I think reverting this change would be the best option. :)

Patch-set: 14