Update patch set 14
Patch Set 14:

> > Re-checked using a full dirty flash. Display is still messed up
 > > like it is when Glove Mode is on.
 > > Further, other Yureka PLUS users are facing the same problem too.
 > >
 > > By the way, I checked via ADB, it shows Normal. What are the
 > other
 > > modes? How do I change them?
 > 
 > Could you try compiling a kernel with the glove mode sysfs node
 > removed, to find out whether this is really the issue here?

Well, I am confused too now. A few other Yureka PLUS users reported today that the latest CM Nightlies are working fine. However, for a few including me, there is this display issue. :/ I know for sure that it is somehow related to Glove Mode because that's how the display reacts but why it is working on a few other PLUS users is the real mystery.

Patch-set: 14