Update patch set 14
Patch Set 14:

Issue Confirmed!
I flashed the 14 September Nightly and noticed it has the same touchscreen issue.
So, I compiled a kernel with no Glove Mode support and then flashed it on 14 September Nightly. It fixed the issue i.e., the touchscreen started responding normally.

So, I urge you to either revert this commit or fix it if you can so that we, Yureka PLUS users, can use Nightlies again. 

In the meantime, I'll release the kernel with NO Glove Mode support to all the Yureka PLUS users. ;)

Patch-set: 14