Update patch set 2
Patch Set 2: -Code-Review

> > > Patch Set 2: Code-Review-1
 > >
 > > Justify reason of "-1" instead of just randomly putting it
 > 
 > I dont think volte should be shipped. You can just keep this patch
 > as it is for others to pick or work on. Furthermore keep proper
 > authorship.

It's your thinking, you aren't supposed to put -1 for your thinking. Why volte can't be shipped? Also, this is a squash commit of few commits, so it's not possible to maintain authorship as patches are from different devs. And "-1" for this is not really a good reason.

Patch-set: 2